Move table elements
