+++ components/Portfolio/index.tsx
@@ -27,55 +27,6 @@ export const SmallTitle = styled.h2`
     white-space: nowrap;
 `;
 
-export const TableHeader = styled.thead``;
-
-export const TableHeading = styled.th`
-    text-align: left;
-    color: var(--color-primary);
-    height: 40px;
-    font-size: var(--font-size-extra-small);
-    padding-left: 10px;
-    border-right: 1px solid var(--color-accent);
-    border-bottom: 1px solid var(--color-accent);
-`;
-
-export const TableLastHeading = styled(TableHeading)`
-    border-right: none;
-`;
-
-export const Table = styled.table``;
-
-export const TableBody = styled.tbody``;
-
-export const TableRow = styled.tr`
-    transition: 0.5s;
-    color: white;
-    opacity: 1;
-
-    &:hover {
-        background: var(--color-accent);
-        cursor: pointer;
-    }
-
-    &.selected {
-        &:hover {
-            background: none;
-            cursor: auto;
-        }
-    }
-`;
-
-export const TableCell = styled.td`
-    padding: 0 10px;
-    border-right: 1px solid var(--color-accent);
-    border-bottom: 1px solid var(--color-accent);
-`;
-
-// Last cell on a table row
-export const TableLastCell = styled(TableCell)`
-    border-right: none;
-`;
-
 export const SecondaryCell = styled.div`
     color: var(--color-secondary);
     font-size: var(--font-size-small);
